backend/pyproject.toml
[tool.poetry]
name = "mlir-modifier-backend"
version = "0.1.0"
description = "Backend for MLIR Modifier - a visual MLIR editing tool"
authors = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-multipart = "^0.0.20"
mlir-python-bindings = "*"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
